requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
xgboost>=1.7.0
lightgbm>=3.3.0
opencv-python>=4.7.0
Pillow>=9.0.0
requests>=2.28.0
nltk>=3.8.0
joblib>=1.2.0
matplotlib>=3.6.0
seaborn>=0.11.0
jupyter>=1.0.0
tqdm>=4.64.0